+++ app1/app.json
@@ -1,9 +1,4 @@
 {
   "name": "app1",
-  "displayName": "app1",
-  "expo": {
-    "android": {
-      "permissions": ["READ_CALENDAR", "WRITE_CALENDAR"]
-    }
-  }
+  "displayName": "app1"
 }
