+++ app1/docs/navigation.md
@@ -1,0 +1,3 @@
+## Refs
+1. https://www.youtube.com/watch?v=UzMbu3XKEoM
+2. https://reactnavigation.org/docs/hiding-tabbar-in-screens
