+++ app1/package.json
@@ -14,9 +14,11 @@
     "format": "eslint -c .eslintrc.js --ignore-path .eslintignore --fix"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^6.3.3",
+    "@react-navigation/native": "^6.0.12",
+    "@react-navigation/native-stack": "^6.8.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "expo": "^46.0.6",
     "expo-asset": "^8.6.1",
     "expo-calendar": "^10.3.0",
     "expo-core": "^3.0.1",
@@ -25,7 +27,9 @@
     "lint-staged": "^13.0.3",
     "pre-commit": "^1.2.2",
     "react": "18.0.0",
-    "react-native": "0.69.3"
+    "react-native": "0.69.3",
+    "react-native-safe-area-context": "^4.3.3",
+    "react-native-screens": "^3.17.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
