+++ app1/src/navigators/index.ts
@@ -1,0 +1,2 @@
+export const navs ="navs"
+export {StackNavigator} from './stack'
