+++ app1/src/screens/index.ts
@@ -1,0 +1,2 @@
+export {CalendarScreen} from './Calendar'
+export {HomeScreen} from './Home'
